move ContainsCIDR to utils package
