+++ pkg/cloudprovider/provider/aws/provider.go
@@ -21,6 +21,7 @@ import (
 	"encoding/json"
 	"errors"
 	"fmt"
+	"github.com/kubermatic/machine-controller/pkg/cloudprovider/util"
 	"strings"
 	"sync"
 	"time"
@@ -796,7 +797,7 @@ func (p *provider) Create(machine *clusterv1alpha1.Machine, data *cloudprovidert
 		},
 	}
 
-	if networkConfig.ContainsCIDR(cloudprovidertypes.IPv6) {
+	if util.ContainsCIDR(networkConfig, util.IPv6) {
 		instanceRequest.NetworkInterfaces[0].Ipv6AddressCount = aws.Int64(1)
 	}
